# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Villa Access Control Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';

export interface AuthResult {
  access_token: string;
  token_type: string;
  user: User;
}